Input control visualization / List control / Add filter support.
Before that, an input control provides all values from the selected
index or all values depending on the parent control selected values.

In a number of situation, the list should only contains a subset of all
possible values. To limit that, adding a filter option which can
contains any query to limit the set of values.

eg. limiting host ending by `.co`
```
+host.keyword:*.co
```

This relies on Elasticsearch filter aggregation.

Related to #16123